Fixed return value bug in input parsing. Fixed grid size calculations bug for concentric circles test case.

+++ src/axom/mir/examples/mir_tutorial_simple.cpp
@@ -97,7 +97,7 @@ struct Input
 
    bool shouldIterate() const { return m_should_iterate; }
    int numIterations() const { return m_iter_count; }
-   int iterPercentage() const { return m_iter_percent; }
+   double iterPercentage() const { return m_iter_percent; }
 
 
    void showhelp()
